requirements.txt
Flask-Mail
itsdangerous
Jinja2
MarkupSafe
Werkzeug>=0.14
wheel
gunicorn
Flask-OAuthlib
Flask-PyMongo
pymongo
dnspython
